+++ graphql/web/http.go
@@ -21,7 +21,6 @@ import (
 	"context"
 	"encoding/json"
 	"io"
-	"io/ioutil"
 	"mime"
 	"net/http"
 	"strings"
@@ -45,8 +44,8 @@ type IServeGraphQL interface {
 	// HTTPHandler returns a http.Handler that serves GraphQL.
 	HTTPHandler() http.Handler
 
-	// Resolver returns a *resolve.RequestResolver that is being used to resolve requests
-	Resolver() *resolve.RequestResolver
+	// Resolve processes a GQL Request using the correct resolver and returns a GQL Response
+	Resolve(ctx context.Context, gqlReq *schema.Request) *schema.Response
 }
 
 type graphqlHandler struct {
@@ -69,8 +68,8 @@ func (gh *graphqlHandler) ServeGQL(resolver *resolve.RequestResolver) {
 	gh.resolver = resolver
 }
 
-func (gh *graphqlHandler) Resolver() *resolve.RequestResolver {
-	return gh.resolver
+func (gh *graphqlHandler) Resolve(ctx context.Context, gqlReq *schema.Request) *schema.Response {
+	return gh.Resolve(ctx, gqlReq)
 }
 
 // write chooses between the http response writer and gzip writer
@@ -106,20 +105,16 @@ func (gh *graphqlHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 	ctx = x.AttachAccessJwt(ctx, r)
 
-	if r.URL.Path == "/admin/schema" {
-		handleAdminSchemaRequest(ctx, w, r, gh)
+	var res *schema.Response
+	gqlReq, err := getRequest(ctx, r)
+
+	if err != nil {
+		res = schema.ErrorResponse(err)
 	} else {
-		var res *schema.Response
-		gqlReq, err := getRequest(ctx, r)
-
-		if err != nil {
-			res = schema.ErrorResponse(err)
-		} else {
-			res = gh.resolver.Resolve(ctx, gqlReq)
-		}
-
-		write(w, res, strings.Contains(r.Header.Get("Accept-Encoding"), "gzip"))
+		res = gh.resolver.Resolve(ctx, gqlReq)
 	}
+
+	write(w, res, strings.Contains(r.Header.Get("Accept-Encoding"), "gzip"))
 }
 
 func (gh *graphqlHandler) isValid() bool {
@@ -190,73 +185,6 @@ func getRequest(ctx context.Context, r *http.Request) (*schema.Request, error) {
 	}
 
 	return gqlReq, nil
-}
-
-// special handler for handling requests to /admin/schema in /alter like fashion
-func handleAdminSchemaRequest(ctx context.Context, w http.ResponseWriter, r *http.Request,
-	gh *graphqlHandler) {
-	x.AddCorsHeaders(w)
-	w.Header().Set("Content-Type", "application/json")
-
-	if r.Method == http.MethodOptions {
-		return
-	} else if !(r.Method == http.MethodPost || r.Method == http.MethodPut) {
-		w.WriteHeader(http.StatusBadRequest)
-		x.SetStatus(w, x.ErrorInvalidMethod, "Invalid method")
-		return
-	}
-
-	if r.Header.Get("Content-Encoding") == "gzip" {
-		zr, err := gzip.NewReader(r.Body)
-		if err != nil {
-			x.SetStatus(w, x.Error, "Unable to create decompressor")
-			return
-		}
-		r.Body = gzreadCloser{zr, r.Body}
-	}
-
-	b, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		x.SetStatus(w, x.ErrorInvalidRequest, err.Error())
-		return
-	}
-
-	gqlReq := &schema.Request{}
-	gqlReq.Query = `
-		mutation updateGqlSchema($sch: String!) {
-			updateGQLSchema(input: {
-				set: {
-					schema: $sch
-				}
-			}) {
-				gqlSchema {
-					id
-				}
-			}
-		}`
-	gqlReq.Variables = map[string]interface{}{
-		"sch": string(b),
-	}
-
-	response := gh.Resolver().Resolve(ctx, gqlReq)
-	if len(response.Errors) > 0 {
-		x.SetStatus(w, x.Error, response.Errors.Error())
-		return
-	}
-
-	res := map[string]interface{}{}
-	data := map[string]interface{}{}
-	data["code"] = x.Success
-	data["message"] = "Done"
-	res["data"] = data
-
-	js, err := json.Marshal(res)
-	if err != nil {
-		x.SetStatus(w, x.Error, err.Error())
-		return
-	}
-
-	_, _ = x.WriteResponse(w, r, js)
 }
 
 func commonHeaders(next http.Handler) http.Handler {
